Default to first `<item>` in a feed when no guid is provided for `{Draper,Feed}Adapter`
When no guid is provided and the `DraperAdapter` or `FeedAdapter` are invoked, they should
return the data as though the guid for the first `<item>` encountered in the feed is provided.

The upshot of this is that it is possible to provide an embed that always points at the latest
episode of a given program, which is pretty handy in a few situations. It's also a reasonably
good default for people who don't care to pick a specific episode and instead want to embed
the "show" itself.