Upgrade to angular 2.4.5, cli beta.26, inline css and js as appropriate
The angular-cli upgrade changes the way that global styles are rendered,
as well as the breakdown of the javascript.

inline.js now has a bundle hash, so we need to search for it. I am now
inlining it because why not.

css for global styles is no longer wrapped in a js file.