Load JS asynchronously in parallel, execute sequentially
This is kind of crummy, but I wasn't able to make it work another way
basically a standard <script> tag will block the DOM including in
certain situations the parent window, making scrolling etc janky.

<script async> works but doesn't guarantee anything about the execution
order of multiple scripts which the angular-cli build depends on.

<script defer> is actually perfect but it's bugged out in IE.

So we do a 2 part hack, with preload tags and a sequential addition
of <script async> tags once each of the previous ones have loaded.

In my testing it made things much less janky and sped up the overall
load