Add basic responsiveness between 150px and 200px high.
Right now, we require 183px, which is fine. I tried a bunch of stuff to make it fluid but the
way the flex is structured right now makes that impossible as near as I can tell, so I just
did it the breakpoint way. Also made the background fill available space.